@@ -19,15 +19,14 @@ package org.apache.camel.k.quarkus.cron.deployment;
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 
 import org.apache.camel.CamelContext;
-import org.apache.camel.ExtendedCamelContext;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.dsl.yaml.YamlRoutesBuilderLoader;
 import org.apache.camel.k.Runtime;
@@ -50,7 +49,7 @@ public class Application {
     @Path("/find-cron-interceptor")
     @Produces(MediaType.TEXT_PLAIN)
     public String findCronInterceptor() {
-        return context.adapt(ExtendedCamelContext.class)
+        return context.getCamelContextExtension()
             .getFactoryFinder(Constants.SOURCE_LOADER_INTERCEPTOR_RESOURCE_PATH)
             .findClass("cron")
             .map(Class::getName)
