@@ -18,22 +18,22 @@ package org.apache.camel.k.quarkus.it;
 
 import java.util.stream.Collectors;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.json.Json;
-import javax.json.JsonObject;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.json.Json;
+import jakarta.json.JsonObject;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import org.apache.camel.CamelContext;
 import org.apache.camel.FluentProducerTemplate;
-import org.apache.camel.model.ModelCamelContext;
+import org.apache.camel.model.Model;
 import org.apache.camel.model.RouteTemplateDefinition;
 
 @RegisterForReflection(targets = { String.class })
@@ -49,7 +49,7 @@ public class Application {
     @Path("/inspect")
     @Produces(MediaType.APPLICATION_JSON)
     public JsonObject inspect() {
-        var templates = context.adapt(ModelCamelContext.class).getRouteTemplateDefinitions();
+        var templates = context.getCamelContextExtension().getContextPlugin(Model.class).getRouteTemplateDefinitions();
         var ids = templates.stream().map(RouteTemplateDefinition::getId).collect(Collectors.toList());
 
         return Json.createObjectBuilder()
