@@ -19,9 +19,9 @@ package org.apache.camel.k.loader.jsh.it;
 import java.io.IOException;
 import java.io.InputStream;
 
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.MediaType;
 
-import io.quarkus.test.junit.DisabledOnNativeImage;
+import io.quarkus.test.junit.DisabledOnIntegrationTest;
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
@@ -29,7 +29,7 @@ import org.apache.camel.util.IOHelper;
 import org.junit.jupiter.api.Test;
 
 //@Disabled
-@DisabledOnNativeImage
+@DisabledOnIntegrationTest
 @QuarkusTest
 public class JshLoaderTest {
     @Test
