@@ -19,9 +19,9 @@ package org.apache.camel.k.loader.kotlin.quarkus;
 import java.io.IOException;
 import java.io.InputStream;
 
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.MediaType;
 
-import io.quarkus.test.junit.DisabledOnNativeImage;
+import io.quarkus.test.junit.DisabledOnIntegrationTest;
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
@@ -32,7 +32,7 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Disabled("https://github.com/quarkusio/quarkus/issues/11549")
-@DisabledOnNativeImage
+@DisabledOnIntegrationTest
 @QuarkusTest
 public class KotlinLoaderTest {
     @Test
