@@ -16,7 +16,7 @@
  */
 package org.apache.camel.k.quarkus.it;
 
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.MediaType;
 
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.path.json.JsonPath;
