@@ -20,22 +20,22 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.json.Json;
-import javax.json.JsonArray;
-import javax.json.JsonObject;
-import javax.json.bind.Jsonb;
-import javax.json.bind.JsonbBuilder;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.json.Json;
+import jakarta.json.JsonArray;
+import jakarta.json.JsonObject;
+import jakarta.json.bind.Jsonb;
+import jakarta.json.bind.JsonbBuilder;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.Route;
-import org.apache.camel.model.ModelCamelContext;
+import org.apache.camel.model.Model;
 import org.apache.camel.model.RouteDefinition;
 import org.apache.camel.model.ToDefinition;
 import org.apache.camel.model.rest.RestDefinition;
@@ -61,11 +61,11 @@ public class RuntimeInspector {
                     .map(Route::getId)
                     .collect(Collectors.toList())))
             .add("route-definitions", Json.createArrayBuilder(
-                camelContext.adapt(ModelCamelContext.class).getRouteDefinitions().stream()
+                camelContext.getCamelContextExtension().getContextPlugin(Model.class).getRouteDefinitions().stream()
                     .map(RouteDefinition::getId)
                     .collect(Collectors.toList())))
             .add("rest-definitions", Json.createArrayBuilder(
-                camelContext.adapt(ModelCamelContext.class).getRestDefinitions().stream()
+                camelContext.getCamelContextExtension().getContextPlugin(Model.class).getRestDefinitions().stream()
                     .map(RestDefinition::getId)
                     .collect(Collectors.toList())))
             .build();
@@ -96,7 +96,7 @@ public class RuntimeInspector {
     @Path("/route-outputs/{name}")
     @Produces(MediaType.APPLICATION_JSON)
     public JsonArray routeOutputs(@PathParam("name") String name) {
-        RouteDefinition def = camelContext.adapt(ModelCamelContext.class).getRouteDefinition(name);
+        RouteDefinition def = camelContext.getCamelContextExtension().getContextPlugin(Model.class).getRouteDefinition(name);
         if (def == null) {
             throw new IllegalArgumentException("RouteDefinition with name: " + name + " not found");
         }
